Revert some env var changes from upstream
* CELERY_BROKER/CELERY_RESULT_BACKEND are back to being supplied
verbatim. This has been changed in upstream to allow urlencoding the
password inside the app, but is tied to the Docker setup, and we
generate the URL inside the NixOS module, so we can go back to providing
the URLs inside env vars directly, which works for us, though isn't more
generic than upstream's solution.

* FLOWER_PORT was deleted from required env vars. The NixOS module
supplies the commandline switches to Flower, and we also use a Unix
socket by default there.