nix: enable running update scripts
Update scripts are moved to `$out/libexec/bookwyrm/updates` and
`update.sh` is patched to use our `manage.py` instead of the `bw-dev`
wrapper.

The update scripts themselves are not patched in a generic manner; it
works for the one script that is there now, but future ones will require
manual review.