Relax `packaging` dep
This is useful for building with poetry2nix, since it pulls in
`packaging` from Nixpkgs for building some of the deps, and `packaging`
in Nixpkgs is ahead of the previous dependency specifier.

It might be more desirable to figure out a way to get the stated version
of `packaging` into the `bookwyrm` package, but this is more expedient
at the moment.